Move InputField files into component folder
